photon/clip-interrogator/requirements.txt
clip-interrogator==0.6.0
Pillow
